Tests/LibWeb/Text/expected/Editing/caret-navigation-sceditor.txt

FAIL: 37 SCEditor caret assertions failed
Down+Up from first list item@1: expected #text@1, got #text@2
Down+Up from first list item@4: expected #text@4, got #text@3
Down+Up from first list item@6: expected #text@6, got #text@5
Down+Up from first list item@9: expected #text@9, got #text@10
Up+Down from second list item@2: expected #text@2, got #text@1
Up+Down from second list item@4: expected #text@4, got #text@3
Up+Down from second list item@6: expected #text@6, got #text@5
Up+Down from second list item@13: expected #text@13, got #text@12
Up+Down from second list item@14: expected #text@14, got #text@12
End from the editing host start: expected DIV@2, got BODY@13
Down from inserted text: expected DIV@0, got #text@3
Up after crossing an empty line: expected #text@1, got #text@0
Shift+End: expected DIV@2, got #text@15
Shift+PageUp: expected #text@0, got #text@5
Shift+PageDown: expected #text@13, got #text@5
Right over first image: expected DIV@2, got DIV@0
Option+Right over first image: expected DIV@2, got #text@15
Option+Left over first image: expected #text@15, got #text@10
Option+Right over punctuation: expected #text@14, got #text@15
Option+Left over punctuation: expected #text@13, got #text@10
Option+Left over trailing punctuation and space: expected #text@13, got #text@10
Option+Right across adjacent styled text: expected #text@4, got #text@13
Option+Left across adjacent styled text: expected #text@13, got #text@0
Option+Shift+Right canonical anchor: expected #text@0, got #text@13
Option+Shift+Right across adjacent styled text: expected #text@4, got #text@13
Shift+Right canonical inline anchor: expected #text@0, got #text@13
Shift+Right across adjacent styled text: expected #text@1, got #text@0
Shift+Down canonical image anchor: expected DIV@1, got #text@15
Shift+Down into formatted block boundary: expected DIV@0, got #text@0
Option+Right across split punctuation: expected #text@1, got #text@10
Option+Left across split punctuation: expected #text@10, got #text@0
Option+Shift+Left across split punctuation: expected #text@0, got #text@0
Option+Left over final image: expected #text@32, got #text@0
Option+Right after final image: expected #text@3, got DIV@4
Option+Shift+Left over final image: expected DIV@3, got #text@32
Up from after the final image: expected DIV@0, got #text@0
Down back to before the final image: expected #text@32, got #text@0
